feat: move states to eliminate validation bugs
add polish
