+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\App.js":"1","C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\contactForm.js":"2","C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\suggestionForm.js":"3","C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\navbar.js":"4","C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\reportWebVitals.js":"5"},{"size":2919,"mtime":1610720001258,"results":"6","hashOfConfig":"7"},{"size":3402,"mtime":1610720559115,"results":"8","hashOfConfig":"7"},{"size":2454,"mtime":1610720999548,"results":"9","hashOfConfig":"7"},{"size":507,"mtime":1610693658131,"results":"10","hashOfConfig":"7"},{"size":362,"mtime":1607397850514,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"w0o91x",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\App.js",["22","23"],"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\contactForm.js",[],"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\suggestionForm.js",[],"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\navbar.js",["24","25"],"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\reportWebVitals.js",[],{"ruleId":"26","severity":1,"message":"27","line":2,"column":26,"nodeType":"28","messageId":"29","endLine":2,"endColumn":35},{"ruleId":"26","severity":1,"message":"30","line":12,"column":10,"nodeType":"28","messageId":"29","endLine":12,"endColumn":19},{"ruleId":"31","severity":1,"message":"32","line":15,"column":17,"nodeType":"33","endLine":15,"endColumn":29},{"ruleId":"31","severity":1,"message":"32","line":16,"column":17,"nodeType":"33","endLine":16,"endColumn":29},"no-unused-vars","'useEffect' is defined but never used.","Identifier","unusedVar","'validated' is assigned a value but never used.","jsx-a11y/anchor-is-valid","The href attribute requires a valid value to be accessible. Provide a valid, navigable address as the href value. If you cannot provide a valid href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md","JSXOpeningElement"]
+[{"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\App.js":"1","C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\contactForm.js":"2","C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\suggestionForm.js":"3","C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\navbar.js":"4","C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\reportWebVitals.js":"5"},{"size":1897,"mtime":1610728705797,"results":"6","hashOfConfig":"7"},{"size":3607,"mtime":1610727806813,"results":"8","hashOfConfig":"7"},{"size":2944,"mtime":1610727590204,"results":"9","hashOfConfig":"7"},{"size":507,"mtime":1610693658131,"results":"10","hashOfConfig":"7"},{"size":362,"mtime":1607397850514,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"w0o91x",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"18","messages":"19","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\App.js",["22"],"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\contactForm.js",[],"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\suggestionForm.js",["23"],"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\components\\navbar.js",["24","25"],"C:\\Users\\isaac\\Desktop\\projects\\MidwayUSA-React\\client\\src\\reportWebVitals.js",[],{"ruleId":"26","severity":1,"message":"27","line":2,"column":26,"nodeType":"28","messageId":"29","endLine":2,"endColumn":35},{"ruleId":"26","severity":1,"message":"30","line":4,"column":12,"nodeType":"28","messageId":"29","endLine":4,"endColumn":21},{"ruleId":"31","severity":1,"message":"32","line":15,"column":17,"nodeType":"33","endLine":15,"endColumn":29},{"ruleId":"31","severity":1,"message":"32","line":16,"column":17,"nodeType":"33","endLine":16,"endColumn":29},"no-unused-vars","'useEffect' is defined but never used.","Identifier","unusedVar","'validated' is assigned a value but never used.","jsx-a11y/anchor-is-valid","The href attribute requires a valid value to be accessible. Provide a valid, navigable address as the href value. If you cannot provide a valid href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md","JSXOpeningElement"]
