+++ client/src/App.css
@@ -8,42 +8,6 @@
   height: 100vh;
 }
 
-/* .App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-} */
-
-
 .navigation {
   background-color: rgba(255, 254, 254, 0.3);
   box-shadow: 0px 2px 4px gray;
@@ -67,14 +31,7 @@
 }
 
 .top #myLinks a {
-  /* padding: 2%;
-  margin: 2%;
   text-decoration: none;
-  border: 1px gray solid;
-  border-radius: 2px;
-  color: #080E33; */
-  text-decoration: none;
-  /* background-color: rgba(6, 22, 53, 0.3); */
   color: rgb(6, 22, 53);
   border-radius: 3px;
   border: none;
@@ -84,14 +41,68 @@
   margin: 1%;
 }
 
-.top #myLinks a:hover {
-  /* background:rgba(6, 22, 53, 0.5); */
-  /* transition: .6s; */
+/* .top #myLinks a:hover {
+  background:rgba(6, 22, 53, 0.5);
+  transition: .6s;
+} */
+
+.toggle {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  margin-top: 1%;
+}
+
+.Button{
+  background-color: rgba(6, 22, 53, 0.1);
+  box-shadow: 0px 2px 4px gray;
+  border-left:2px #C0C0C0 solid;
+  border-right:2px #C0C0C0 solid;
+  border-top:2px #C0C0C0 solid;
+  border-bottom: none;
+  border-radius: 3px;
+  float: center;
+  cursor: pointer;
+  padding: 14px 16px;
+  transition: 0.3s;
+  width: 100%;
+  font-size: 110%;
+  color: rgb(6, 22, 53);
+}
+
+.Button:hover{
+  background:rgba(6, 22, 53, 0.3);
+}
+
+.active{
+  background: rgba(255, 255, 255, 0.4);
+  color: rgb(6, 22, 53);
+  border: none;
+  outline: none;
+}
+.active:hover{
+  background:rgba(6, 22, 53, 0.3);
+  color: rgb(6, 22, 53);
+}
+
+
+.tab {
+  overflow: hidden;
+  width: 39.5%;
+  display: flex;
+  justify-content: space-around;
+}
+
+.activeTab{
+  display: block;
+}
+  
+.tabContent{
+  display: none;
 }
 
 .container {
   display: flex;
-  /* flex-direction: column; */
   justify-content: center;
   text-align: center;
   width: 100%;
@@ -100,18 +111,20 @@
 .middle {
   width: 40%;
   height: 50vh;
+  margin-top: 0;
+}
+
+.formCard{
+  width: 100%;
+  padding: 2% 0;
 }
 
 .productSuggestionForm, .contactUsForm {
-  /* border: 1px red solid; */
   width: 100%;
 }
 
-
-
 .productForm, .contactForm {
-  /* display: none; */
-  background: rgba(255, 255, 255, 0.3);
+  background-color: rgba(6, 22, 53, 0.1);
   box-shadow: 0px 2px 4px gray;
   margin-top: 2%;
   padding-top: 2%;
@@ -131,7 +144,7 @@
   width: 80%;
   max-width: 80%;
   min-width: 80%;
-  background-color: rgba(6, 22, 53, 0.1);
+  background-color: rgba(252, 252, 252, 0.4);
 }
 
 .field select{
@@ -144,7 +157,7 @@
   border-top-right-radius: 3px;
   border-top-left-radius: 3px;
   width: 81%;
-  background-color: rgba(6, 22, 53, 0.2);
+  background-color: rgba(252, 252, 252, 0.4);
 }
 
 .actions input {
@@ -160,13 +173,15 @@
 
 .actions input:hover {
   background:rgba(6, 22, 53, 0.8);
-  /* transition: .6s; */
 }
 
 .messages {
   text-align: center;
 }
 
+.messages p{
+  color: red;
+}
 
 footer {
   text-align: center;
@@ -174,64 +189,3 @@ footer {
   position: fixed;
   width: 100%;
 }
-
-.toggle {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  margin-top: 1%;
-}
-
-.Button{
-  /* background-color: inherit;
-  float: center;
-  border: none;
-  outline: none;
-  cursor: pointer;
-  padding: 14px 16px;
-  transition: 0.3s;
-  width: 100%;
-  font-size: 2rem;
-  color: #470B97; */
-
-  background-color: rgb(6, 22, 53);
-  color: white;
-  border-radius: 3px;
-  border: none;
-  font-size: 1rem;
-  padding: 2% 1%;
-  width: 40%;
-  margin: 1%;
-  
-  
-}
-.Button:hover{
-  background:rgba(6, 22, 53, 0.8);
-}
-
-.tab {
-  overflow: hidden;
-  /* background-color:  #C0C0C0; */
-  width: 39.5%;
-  display: flex;
-  justify-content: space-around;
-}
-
-.activeTab{
-  display: block;
-}
-  
-.tabContent{
-  display: none;
-}
-  
-
-  
-.loginCard{
-    /* background-color: #C0C0C0;
-    border-top: 1px #470B97 solid;
-    border-bottom: 1px #470B97 solid; */
-    width: 100%;
-    padding: 2% 0;
-}
-
